canSum

+++ src/base/dynamicProgramming/FibNthNumber.java
@@ -7,7 +7,7 @@ import java.util.Map;
 
 public class FibNthNumber {
     public static void main(String[] args) {
-        Map<Integer,Integer> map=new HashMap<Integer, Integer>() ;
+        Map<Integer,Long> map=new HashMap<Integer, Long>() ;
         SimpleDateFormat sdf= new SimpleDateFormat("dd-MM-yyyy hh:mm:ss.MS");
         Date date1= new Date();
         System.out.println(sdf.format(date1));
@@ -20,19 +20,19 @@ public class FibNthNumber {
         System.out.println(sec);
     }
 
-    public static int  fib_recurr(int n){ // for n=50 it took 58s
+    public static Long  fib_recurr(int n){ // for n=50 it took 58s
         if(n<=2)
-            return 1;
+            return (long)1;
         else
             return (fib_recurr(n-1) + fib_recurr(n-2));
     }
 
     //storing the sub-problem to reduce the time.This is Dynamic programming
-    public static int  fib_dyn(int n, Map<Integer,Integer> map){// for n=50 it took 0secs
+    public static Long  fib_dyn(int n, Map<Integer,Long> map){// for n=50 it took 0secs
         if(map.containsKey(n))
             return map.get(n);
         else if(n<=2)
-            return 1;
+            return (long)1;
         else
         {
             map.put (n,fib_dyn(n-1,map) + fib_dyn(n-2,map));
